Fix migrations, evitando que quebre produção

+++ db/migrate/20130709121904_remove_relationship_between_expense_natures_and_regulatory_acts.rb
@@ -1,5 +1,13 @@
 class RemoveRelationshipBetweenExpenseNaturesAndRegulatoryActs < ActiveRecord::Migration
   def change
-    remove_column :compras_expense_natures, :regulatory_act_id
+    if connection.table_exists? :compras_expense_natures
+      remove_column :compras_expense_natures, :regulatory_act_id
+    end
+  end
+
+  protected
+
+  def connection
+    ActiveRecord::Base.connection
   end
 end
