+++ db/migrate/20130709142353_remove_relationship_between_agreements_and_regulatory_acts.rb
@@ -1,5 +1,13 @@
 class RemoveRelationshipBetweenAgreementsAndRegulatoryActs < ActiveRecord::Migration
   def change
-    remove_column :compras_agreements, :regulatory_act_id
+    if connection.table_exists? :compras_expense_natures
+      remove_column :compras_agreements, :regulatory_act_id
+    end
+  end
+
+  protected
+
+  def connection
+    ActiveRecord::Base.connection
   end
 end
