+++ db/migrate/20130711170227_remove_administration_types.rb
@@ -1,6 +1,17 @@
 class RemoveAdministrationTypes < ActiveRecord::Migration
   def change
-    remove_column :compras_budget_structures, :administration_type_id
-    drop_table :compras_administration_types
+    if connection.table_exists? :compras_budget_structures
+      remove_column :compras_budget_structures, :administration_type_id
+    end
+
+    if connection.table_exists? :compras_administration_types
+      drop_table :compras_administration_types
+    end
+  end
+
+  protected
+
+  def connection
+    ActiveRecord::Base.connection
   end
 end
