+++ db/migrate/20130718174132_remove_relationship_between_creditor_and_precatories.rb
@@ -1,5 +1,13 @@
 class RemoveRelationshipBetweenCreditorAndPrecatories < ActiveRecord::Migration
   def change
-    remove_column :compras_precatories, :creditor_id
+    if connection.table_exists? :compras_precatories
+      remove_column :compras_precatories, :creditor_id
+    end
+  end
+
+  protected
+
+  def connection
+    ActiveRecord::Base.connection
   end
 end
