+++ db/migrate/20130718175422_remove_relationship_between_creditor_and_reserve_fund.rb
@@ -1,5 +1,13 @@
 class RemoveRelationshipBetweenCreditorAndReserveFund < ActiveRecord::Migration
   def change
-    remove_column :compras_reserve_funds, :creditor_id
+    if connection.table_exists? :compras_reserve_funds
+      remove_column :compras_reserve_funds, :creditor_id
+    end
+  end
+
+  protected
+
+  def connection
+    ActiveRecord::Base.connection
   end
 end
